requirements.txt
requests~=2.32.3
python-socketio~=5.13.0
chess~=1.11.2
stockfish~=3.28.0
fastapi~=0.115.12
pydantic~=2.11.4
sqlmodel~=0.0.24